// Frontend/src/app/api/models/public-job-offer-list-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface PublicJobOfferListDto {
  applicationCount?: number;
  dateLimite?: string;
  datePublication?: string;
  description?: string;
  experienceMin?: number;
  id?: number;
  jobTitle?: string;
  jobType?: 'PER' | 'PATS' | 'CONTRACTUEL';
  niveauEtudeRequis?: string;
  requiredSkills?: Array<string>;
  typeContrat?: 'CDD' | 'CDI' | 'STAGE';
}